Finish tp6 correction and do some improvements in tp code

+++ tp1/src/com/jb/servlets/ListeCommandes.java
@@ -1,8 +1,6 @@
 package com.jb.servlets;
 
 import java.io.IOException;
-import java.util.Date;
-import java.util.HashMap;
 import java.util.Map;
 
 import javax.servlet.ServletException;
@@ -11,8 +9,8 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
 
-import com.jb.beans.Customer;
 import com.jb.beans.Order;
+import com.jb.dao.DAOFactory;
 import com.jb.dao.OrderDao;
 
 public class ListeCommandes extends HttpServlet {
@@ -26,42 +24,13 @@ public class ListeCommandes extends HttpServlet {
 
 	private OrderDao orderDao;
 
+	public void init() throws ServletException {
+		/* Récupération d'une instance de notre DAO Utilisateur */
+		DAOFactory factory = (DAOFactory) getServletContext().getAttribute(CONF_DAO_FACTORY);
+		this.orderDao = factory.getOrderDao();
+	}
+
 	public void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-		/* Affichage de la page d'inscription */
-		HttpSession session = request.getSession();
-		Map<Long, Order> orderList = (Map<Long, Order>) session.getAttribute(ATT_SESSION_ORDERS);
-		if (orderList == null) {
-			orderList = new HashMap<Long, Order>();
-
-			Map<Long, Customer> customerList = (Map<Long, Customer>) session.getAttribute(ATT_SESSION_CUSTOMERS);
-			if (customerList == null)
-				customerList = new HashMap<Long, Customer>();
-
-			for (int i = 1; i < 10; i++) {
-				Customer c = new Customer();
-				String lastName = "last Name " + i;
-				c.setId((long) i);
-				c.setLastName(lastName);
-				c.setFirstName("first Name " + i);
-				c.setAddress("address " + i);
-				c.setPhone("100000" + i);
-				c.setMail("aaa" + i + "@test.com");
-				customerList.put((long) i, c);
-				Order o = new Order();
-				o.setId((long) i);
-				o.setDate(new Date(2016, 4, i));
-				o.setCustomer(c);
-				o.setAmount(100.0 + i);
-				o.setPaymentMethod("aa" + i);
-				o.setPaymentStatus("bb" + i);
-				o.setDeliveryMode("cc" + i);
-				o.setDeliveryStatus("dd" + i);
-				orderList.put((long) i, o);
-			}
-			session.setAttribute(ATT_SESSION_ORDERS, orderList);
-			session.setAttribute(ATT_SESSION_CUSTOMERS, customerList);
-
-		}
 		this.getServletContext().getRequestDispatcher(VIEW).forward(request, response);
 	}
 
@@ -77,7 +46,7 @@ public class ListeCommandes extends HttpServlet {
 
 			HttpSession session = request.getSession();
 			Map<Long, Order> orderList = (Map<Long, Order>) session.getAttribute(ATT_SESSION_ORDERS);
-			if (orderList.containsKey(order)) {
+			if (orderList != null && order != null && orderList.containsKey(order)) {
 				orderList.remove(order);
 				session.setAttribute(ATT_SESSION_ORDERS, orderList);
 				orderDao.delete(order);
